refactor(ui): TE-1371 make routing for create and update use same component
